+++ lore-server/src/grpc/server.rs
@@ -45,6 +45,7 @@ use super::lock_service::LoreLockService;
 use crate::auth::jwt::JwtVerifier;
 use crate::auth::jwt_interceptor::JWTInterceptor;
 use crate::authnz::repository_authorizer::RepositoryAuthorizer;
+use crate::authnz::repository_catalog::RepositoryCatalog;
 use crate::correlation::layer::CorrelationIdLayer;
 use crate::correlation::layer::CorrelationIdLayerBuilder;
 use crate::correlation::layer::TraceLayerConfig;
@@ -633,6 +634,7 @@ impl GrpcServerBuilder<MaybeJwtVerifier> {
         self,
         jwt_verifier: Option<JwtVerifier>,
         repository_authorizer: Arc<dyn RepositoryAuthorizer>,
+        repository_catalog: Arc<dyn RepositoryCatalog>,
     ) -> Result<GrpcServerBuilder<WantsAddress>> {
         let rpc_timeout = self.0.request_handler_timeout;
         let services = &self.0.service_settings;
@@ -732,6 +734,7 @@ impl GrpcServerBuilder<MaybeJwtVerifier> {
         let repository_svc = LoreRepositoryService::new(
             self.0.environment.clone(),
             repository_authorizer.clone(),
+            repository_catalog.clone(),
             self.0.immutable_store.clone(),
             self.0.mutable_store.clone(),
             self.0.hook_dispatcher.clone(),
@@ -740,6 +743,7 @@ impl GrpcServerBuilder<MaybeJwtVerifier> {
         let repository_v1_svc = LoreRepositoryV1Service::new(
             self.0.environment.clone(),
             repository_authorizer.clone(),
+            repository_catalog,
             self.0.immutable_store.clone(),
             self.0.mutable_store.clone(),
             self.0.hook_dispatcher.clone(),